+++ backend/canisters/user/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `set_message_reminder` ([#3417](https://github.com/open-ic/open-chat/pull/3417))
 - Implement 'Gated Groups' ([#3406](https://github.com/open-ic/open-chat/pull/3406))
 - Added `Empty` event type ([#3439](https://github.com/open-ic/open-chat/pull/3439))
+- Added new message content types for reminders ([#3440](https://github.com/open-ic/open-chat/pull/3440))
 
 ### Changed